codigo em java

+++ src/br/com/dio/calculadora/calculadora.java
@@ -15,7 +15,7 @@ public class calculadora {
         int somar = somar(a,b);
         int subtrair = subtrair(a,b);
         int multiplicar = multiplicar(a,b);
-        int dividir = dividir(a,b);
+        double dividir = dividir(a,b);
 
         System.out.println("somar: " + somar);
         System.out.println("subt: " + subtrair);
@@ -31,7 +31,7 @@ public class calculadora {
     public static int multiplicar(int a, int b){
         return a * b;
     }
-    public static int dividir(int a, int b){
+    public static double dividir(double a, double b){
         return a / b;
     }
